SearchActivity: Simplify ArrayList data management
Instead of making a new ArrayList when we parse the JSON data, let's make the ArrayList at the beginning, add the items to it, then tell the adapter to refresh itself. I'm also clearing out some now-unused method parameters for the sake of simplicity. For example, if we don't need to pass the Context around, just don't.

This will change in the future though, as I eventually plan to move the networking and data parsing functionality to a separate class.